// src/store/types/vendor.actionTypes.ts
export enum VENDOR_ACTION_TYPES {
    
  // Create Vendor
  CREATE_VENDOR_REQUEST = 'CREATE_VENDOR_REQUEST',
  CREATE_VENDOR_SUCCESS = 'CREATE_VENDOR_SUCCESS',
  CREATE_VENDOR_FAILURE = 'CREATE_VENDOR_FAILURE',

  // Get All Vendors
  GET_ALL_VENDORS_REQUEST = 'GET_ALL_VENDORS_REQUEST',
  GET_ALL_VENDORS_SUCCESS = 'GET_ALL_VENDORS_SUCCESS',
  GET_ALL_VENDORS_FAILURE = 'GET_ALL_VENDORS_FAILURE',

  // Get Vendor by ID
  GET_VENDOR_BY_ID_REQUEST = 'GET_VENDOR_BY_ID_REQUEST',
  GET_VENDOR_BY_ID_SUCCESS = 'GET_VENDOR_BY_ID_SUCCESS',
  GET_VENDOR_BY_ID_FAILURE = 'GET_VENDOR_BY_ID_FAILURE',

  // Update Vendor
  UPDATE_VENDOR_REQUEST = 'UPDATE_VENDOR_REQUEST',
  UPDATE_VENDOR_SUCCESS = 'UPDATE_VENDOR_SUCCESS',
  UPDATE_VENDOR_FAILURE = 'UPDATE_VENDOR_FAILURE',

  // Delete Vendor
  DELETE_VENDOR_REQUEST = 'DELETE_VENDOR_REQUEST',
  DELETE_VENDOR_SUCCESS = 'DELETE_VENDOR_SUCCESS',
  DELETE_VENDOR_FAILURE = 'DELETE_VENDOR_FAILURE',

  // Update Vendor Stats
  UPDATE_VENDOR_STATS_REQUEST = 'UPDATE_VENDOR_STATS_REQUEST',
  UPDATE_VENDOR_STATS_SUCCESS = 'UPDATE_VENDOR_STATS_SUCCESS',
  UPDATE_VENDOR_STATS_FAILURE = 'UPDATE_VENDOR_STATS_FAILURE',

  // Clear Errors
  CLEAR_VENDOR_ERRORS = 'CLEAR_VENDOR_ERRORS',
}